+++ Client/Client.Qt/Widgets/CaptionButton.hpp
@@ -5,14 +5,29 @@
 
 #include "Style/StyleBasic.hpp"
 
+/**
+ * @class CaptionButton
+ * @brief A button for title
+ */
 class CaptionButton : public AbstractButton
 {
     Q_OBJECT
     Q_PROPERTY(QColor _hoverColor READ getHoverColor WRITE setHoverColor)
 public:
+    /**
+     * @brief Constructor for caption button
+     * @param parent Title widget
+     * @param endColor Color on hover
+     * @param icon Default icon
+     */
     explicit CaptionButton(QWidget* parent = nullptr, const QColor& endColor = QColor(255, 255, 255, 76),
                            const Style::icon& icon = Style::icon());
 
+    /**
+     * @brief Override icon by a new one
+     * @param icon New icon
+     * @return Icon has been set?
+     */
     bool setIcon(const Style::icon* icon);
 
 protected:
