+++ Client/Client.Qt/Widgets/TitleWidget.cpp
