+++ Client/Client.Qt/Widgets/TitleWidget.hpp
